rename: Lotto 클래스를 도메인 패키지로 이동 및 Application 클래스를 상위 패키지로 이동
