+++ src/main/java/Application.java
@@ -1,7 +1,6 @@
-package lotto;
 
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+
     }
 }
