+++ src/main/java/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package domain;
 
 import static utils.NumberValidation.*;
