+++ src/test/java/ApplicationTest.java
@@ -1,5 +1,3 @@
-package lotto;
-
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
