+++ src/test/java/domain/LottoTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package domain;
 
+import domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
